refactor(MatchInfo): MatchRecord와의 연관관계에서 cascade, orphanRemoval 설정 추가
