+++ src/main/java/com/geonwoo/solokill/domain/matchInfo/model/MatchInfo.java
@@ -1,8 +1,11 @@
-package com.geonwoo.solokill.domain.matchrecord.model;
+package com.geonwoo.solokill.domain.matchInfo.model;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import com.geonwoo.solokill.domain.matchrecord.model.MatchRecord;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
@@ -15,7 +18,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MatchInfo {
 
-	@OneToMany(mappedBy = "matchInfo")
+	@OneToMany(mappedBy = "matchInfo", cascade = {CascadeType.PERSIST, CascadeType.REMOVE}, orphanRemoval = true)
 	private final List<MatchRecord> matchRecords = new ArrayList<>();
 
 	@Id
