+++ src/main/java/com/geonwoo/solokill/domain/matchrecord/repository/MatchInfoRepository.java
@@ -2,7 +2,7 @@ package com.geonwoo.solokill.domain.matchrecord.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.geonwoo.solokill.domain.matchrecord.model.MatchInfo;
+import com.geonwoo.solokill.domain.matchInfo.model.MatchInfo;
 
 public interface MatchInfoRepository extends JpaRepository<MatchInfo, String> {
 }
